Update basic_canada.py
I've added more information in the header, to match what the BMW Remote App is sending. The main thing that I think makes it work is the "Authorization" field. This string of numbers and letters seems to be hard coded into the BMW app, and if you don't send it the server will send a empty response. 

Also fixed the VEHICLES_URL and changed response.status.code to 200, as the server sends a HTTP 200 code when the authentication is successful.

With these changes I can successfully get a response with the information of my car.